misc: Fix trivial compiler warnings
Include stdlib.h for abort, free and string for memset, make the
argument const.

+++ platforms/unix/plugins/DropPlugin/sqUnixDragDrop.c
@@ -71,7 +71,7 @@ int dropRequestFileHandle(int dropIndex)
     {
       // you cannot be serious?
       int handle= instantiateClassindexableSize(classByteArray(), fileRecordSize());
-      sqFileOpen((SQFile *)fileValueOf(handle), (int)path, strlen(path), 0);
+      sqFileOpen((SQFile *)fileValueOf(handle), path, strlen(path), 0);
       return handle;
     }  
   return interpreterProxy->nilObject();
